Don't force unset fields to null when clearing fields
I noticed that a bunch of my messages had a bunch of trailing nulls for their
serialized form. I traced this down to oneof fields forcefully setting all the
fields to null, even if they're beyond the length of the backing data array
(and thus thereby increased the size of the serialized form).

We can instead simply delete the value. If it was set it will now be unset, and if it was never set then it's a noop.

PiperOrigin-RevId: 309479952