Shift the inferred pivot by one if there isn't a message id
If there isn't a message ID (i.e. we're using 0-indexed serialization) then we
will end up inferring the pivot to be at the length of the data. However, that
means with 0-indexed serialization, field number `data.length` is actually set
at index `data.length - 1`. This is problematic since we'll now assume that
field number `data.length` is contained within the extension object when it
actually isn't.

This bug doesn't reproduce with 1-indexed serialization since we don't end up
having to shift the index.

PiperOrigin-RevId: 310214165