Remove explicit dependency to JRE.
This is no longer required as bootclasspath includes all JRE now.

It might be the case that JavaScript code might depend on a J2CL proto target without the app depending on any other J2CL code. In that case the app will miss needed transpiled JRE but such a scenario is extremely unlikely.

PiperOrigin-RevId: 322851040