Use code generation for `JspbKernel`/`InternalMessage`
Many of the methods on these classes are repetitive and would be better served
by using a generator to avoid subtle mistakes in a sea of similar method
implementations.

To enable this a few minor changes were made to `FieldAccessor` to add `Int52`
specific field getters. Previously they were manually mapped to the `Long`
variants, but explicitly doing this in the `FieldAccessor` avoids the need to
special-case the codegen.

In addition, I exposed some types from the existing implementation codegen so
that they could be shared with the runtime codegen. Namely this meant exposing
more `TypeDescriptor` so that they could be reused in both code generators.

Notably I didn't codegen `InternalMessage.ExtensionFieldInfo` since there's a
discrepancy between Float/Double. I'll work on resolving this and adding codegen
for that in a follow-up.

PiperOrigin-RevId: 332911523