Introduce a InternalMutableListView to provide a write-API to InternalMessage
This change allows us to remove many of the type-specific methods from
JspbKernel as we collapse everything down to `getRepeated{Stem}Field` and
`getMutableRepeated{Stem}Field`. InternalMessage can then use those API to
directly access the values rather than have to be proxied through JspbKernel.

This did require making the setterFn on InternalList lazy as well so that we
could remove the setterFn from the InternalMutableListView API, allowing that
interface to be kernel agnostic.

The only list releated method that has special treament is the count method.
Since that method doesn't need to be type specific (reading the length of the
array doesn't require an accessorFn), I had to leave it as its own generic
method on the kernel.

PiperOrigin-RevId: 335943133