Create internal components for protobuf Map representations
This change defines the base components for Map support. The foundation is the
MapView/InternalMap which define a read-only API and a mutable wire-format aware
view of the underlying data, respectively.

To support this we needed to define FieldAccessors for working with the
underlying JSPB data, largely just wraps existing FieldAccessors in order to
read/set underlying values.

PiperOrigin-RevId: 336918452