Generate message API methods for Map fields
With this Map fields will now have the map API available for use, rather than
having to treat them as a repeated submessage. The API methods being added are:

On Message/Builder:

  - get{Field}Map()
  - get{Field}OrDefault(key. defaultValue)
  - get{Field}OrThrow(key)
  - containsField(key)
  - get{Field}Count()

On Builder:

  - put{Field}(key, value)
  - putAll{Field}(entries)
  - remove{Field}(key)

Note: get{Field}Count() did previously exist for map fields as it was simply
treated as a repeated field. While the method signature is the same, the
returned value may now be different. In particular, previously the value
returned were the number of elements in the repeated field, whereas we now
return the number of unique keys in the field. Unless there was a malformed map
(i.e. duplicates keys), you shouldn't see a change in this value.
PiperOrigin-RevId: 339365448