Implement the Map API for J2CL proto
This wraps the existing Immutable JSPB Map API with a small shim to adapt a
MapView into a Java Map.

For now the repeated field view on map fields is left in place until all users
can be migrated off of that. The only consequence of this is that there's
overlap on the `Count()` method. To maintain behavior the repeated field version
is left in place, which may produce a different result than the map version
would if there are repeated keys in the payload.

PiperOrigin-RevId: 340971689