Fix edge where extension storage accessed after Builder.build.
Extension storage (the storage after pivot) is lazily initialized. That means if the storage wasn't accessed by Builder, then the storage might get initialized after the build on field access.

Existing build method is calling freeze after the copy to catch potential implementation issues and this resulted in exception to be thrown in scenario that is explained earlier.

Since freezing is not recursive anyway, we are not really catching more interesting accidental modifications so I don't see much value in trying to workaround it. Hence I removed it instead.

Alternatives considered:
1. Eagerly initialize array.
This will result in all storage to be initialized up to the pivot point which will increase memory usage and wire size unnecessarily.

2. Update hasser/getters to pass a flag to not initialize storage field.
This introduces extra complexity and also has potential implication on builder API that may eagerly create a wrapper object (ListView/MapView).

3. Check and silently return empty array from getStorageFor.
I don't like this option since it will result in silent mistakes if the storage object is accessed and object is array was frozen accidentally. This will defeat one of the main point of freezing and harder to debug the issue.

PiperOrigin-RevId: 341868341