Rework pivot selection to be based on field number rather than index
Largely this required reworking getStorageFor to work based on the field number rather than the adjusted index since we otherwise don't know if the adjusted index is for the array or the extension object.

Additionally, if a message has no fields, assume the max field number is 1. If the message is extendable, use that as the suggested pivot.

PiperOrigin-RevId: 398504742