Rework createGetDescriptorFnFromArgs to lazily get the args
Because we can refer to other descriptors in the same file we want to avoid
immediately referring to them as the file is loaded. Some of the descriptors
being referenced may not yet be defined. To avoid having to work through the
exact ordering we can instead pass the args through an arrow function and
execute it the first time the getDescriptor is called.

While I was here I also exported the ExtensionRegistry type.

PiperOrigin-RevId: 413145979