Rename internal descriptor namespaces
Before I was putting everything on the `proto.im.internal` namespace which is
not great. I think better matching the directory structure makes this easier to
understand. The only one that is a bit wonky is
`proto.im.descriptor.internal.descriptor`, but I think it's tolerable,
especially since it's not intended for public consumption.

PiperOrigin-RevId: 420767443