Rework the descriptor testUtils into (largely) a builder pattern
For most cases we can use a simpler builder pattern to construct Descriptors for
testing. There's a few cases where we want to generate malformed descriptors to
test with, but that's the minority of cases.

While I was here I also moved the utility class into a proper testing directory.

PiperOrigin-RevId: 420771335