Render extensions after all messages/enums have been rendered
This avoids an order issue where we attempt to create extensions that refer to a
type not yet defined. Currently we end up passing in an undefined ctor function
for these extensions would eventually later fail at runtime.

I also added a guard to ensure that we're not passing in undefined ctor
functions when we construct extensions.

PiperOrigin-RevId: 423948864