Roll forward of https://github.com/google/j2cl-proto/commit/e5b37064a1933f8bd12a19c601a36caf88bb1569: Generate descriptors for ImmutableJS messages
For now these should be entirely unused so there shouldn't be any code size
impact from this. Messages will expose a public `getDescriptor` static method
that can be used to reference the relevant descriptor for that particular
message type.

I had to make a few tweaks to how extensions are created in order to ensure that
they can be fully pruned if never referenced, namely this means that we need to
directly assign the extension onto the registry. If we go through the
`registerExtension` helper then the registry will be considered read and thus no
longer prunable.

NEW: Fixed an ordering issue where extensions would attempt to register
themselves before the parent message descriptor had been created.
PiperOrigin-RevId: 424094752