Write unsigned 64-bit integers as their scale rather than bit representation
This requires creating a new type to represent unsigned longs to ensure that we
write the unsigned scale to the wire. On the read side nothing is changing as
we're still representing the values in their signed representation. The setters
for these fields will now accept both signed and unsigned representations of the
values as well.

If the field is not using string representation for the value then we're just
going to use the normal int52 representation that exists today since we're
already far from being able to use the sign bit there anyway.

PiperOrigin-RevId: 425623589