Ensure that identity equality is maintained for default messages
Currently when you attempt to get an unset message we're returning a new empty
instance instead of the default instance.

Since this only impacts singular message fields, we can simply perform a
hasField() check before either returning the real default instance, or read the
underlying data.

Repeated/map values are not impacted by this as we do not return default
instances if you try to read out of bounds or unset keys.

This does make reading singular map fields slightly more expensive as we're
performing the extra check, but it's probably marginal.

PiperOrigin-RevId: 454673906