Fix handling of unrecognized enum values.
When proto2 message refers proto3 enums, the behavior is similar to proto2 enums; unknown values are treated as default. However J2CL was handling them as unrecognized. This fix addresses that.

I will follow up with furter refactoring for the WASM part which uses "getDefaultValue" code path and this CL avoids touching that part for now.

PiperOrigin-RevId: 463269918