Relax field equivalence checks to handle more variations in wire format
A number field can either be encoded as a number literal, or as a string. There
are a few special values that are always encoded as a String: `'NaN'`,
`'Infinity'`, `'-Infinity'`. The latter two can be handled by JS's coerced
equality, however, `NaN != 'NaN'` (since `Nan != NaN`) and thus we need
additional checks for this case.

A bool field can be present on the wire as either a boolean literal or as a
number. If it's a number then only 0 is treated as false.

This does mean that this solution is not complete as I require one of the values
being compared to be a boolean. That is, two bool fields encoded with non-zero
numbers (ex. a = 2 and b = 3) should be treated as equal. However, we do not
have field type information available to us and cannot perform this check.

PiperOrigin-RevId: 475388122